\name{conc133}
\alias{conc133}
\docType{data}
\title{Concentration 95}
\description{
  This datasets contains the values of the latine square matrix for the hgu133 Affymetrix Microarrays
}
\usage{data(conc133)}
\format{
  The format is:
 num [1:14] 0 0.125 0.25 0.5 1 2 4 8 16 32 ...
}
\seealso{
\code{\link{Ivsc}}, \code{\link{IvsDG}}, \code{\link{collapse}}, \code{\link{SPIKE_IN}}, \code{\link{SPIKE_INA}}, \code{\link{SPIKE_INB}}, \code{\link{SPIKE_INH}}
}
\examples{
data(conc133)
}
\keyword{datasets}
