\name{conc95}
\alias{conc95}
\docType{data}
\title{Concentration 95}
\description{
  This datasets contains the values of the latine square matrix for the hgu95 Affymetrix Microarrays
}
\usage{data(conc95)}
\format{
  The format is:
 num [1:14] 0 0.25 0.5 1 2 4 8 16 32 64 ...
}
\seealso{
\code{\link{Ivsc}}, \code{\link{IvsDG}}, \code{\link{collapse}}, \code{\link{SPIKE_IN}}, \code{\link{SPIKE_INA}}, \code{\link{SPIKE_INB}}, \code{\link{SPIKE_INH}}
}
\examples{
data(conc95)
}
\keyword{datasets}
